refactor: Move DM modules to modules directory
