+++ application/millage-client/src/modules/DM/reducer.tsx
@@ -19,4 +19,4 @@ const DMReducer = createReducer<DMState, DMAction>(initialState, {
   }),
 });
 
-export {DMReducer};
+export default DMReducer;
